Don't force BP2 to be the front application before displaying an alert or file dialog.
CallUser() was using SetFrontProcess() to make BP2 the active application before it called AEInteractWithUser() which made the latter call irrelevant.  AEInteractWithUser() is a better way to ask the user to activate BP2: it beeps and bounces the icon in the dock for 40 seconds before canceling the alert function that called CallUser().