Conditionalized SaveLoads3.c and started modifying some routines for console build.
Factored out prefix/extension matching from IdentifyBPFileType()
into a new function IdentifyBPFileTypeByName() that may be
called with just a filename.  This and some other functions
that are now available in SaveLoads3.c will be useful in
argument parsing for the console version.

WriteToFile() and NoReturnWriteToFile() now should work in 
the console build but need further abstraction in their
parameters.

Created stub for CheckTextSize().
Other minor fixes for old code.