Conditionalized parts of LoadScriptCommands() which may not be needed.  See comments below...
The list of script commands is only used in a few isolated 
places in code which is included in the console build. 
CompileGlossary() and IsEmpty() may crash because they use
the macro p_ScriptLabelPart(). (DisplayGrammar() does also 
but it currently is not called anywhere).

The space for the ScriptLine global does need to be allocated
in the console build as many functions try to append script lines.