Replaced uses of NewHandle() and DisposeHandle() in Inits.c and GetRelease.c.
I'm not sure why NewHandle() was being used in this section of
Inits(), possibly because it is before the call to MakeWindows()
and there was some possibility that GiveSpace() would make an
error call that needed the windows initialized?  However, 
LoadStrings() is called just before this section and causes 
GiveSpace() to be called.

Another possibility might be to avoid counting this memory
in the totals for leak detection (although there was some
possible misaccounting then here and in ClearLockedSpace()) ?